requirements.txt
image2text == 0.1.0
wordcloud == 1.5.0

# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# server/__init__.py: 1
# server/views.py: 5
Flask == 1.0.2

# mkcloud/main.py: 4
Pillow == 5.2.0

# mkcloud/main.py: 5
numpy == 1.15.0

# summarize-test.py: 1
# wordcloud-test.py: 1
requests == 2.19.1
